requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.5.0

# Optional - for visualization tools
pillow>=8.0.0

